+++ bosesoundtouchapi/models/preset.py
@@ -56,6 +56,9 @@ class Preset(SoundTouchModelRequest):
         self._UpdatedOn:int = 0
         self._ContentItem:ContentItem = ContentItem()
 
+        # helper properties (non-xml).
+        self._SourceTitle:str = None
+
         if (root is None):
             
             if isinstance(source, SoundTouchSources):
@@ -205,6 +208,25 @@ class Preset(SoundTouchModelRequest):
 
 
     @property
+    def SourceTitle(self) -> str:
+        """ 
+        The source title of media content (e.g. "Tunein", "Airplay", "NAS Music Server", etc). 
+        
+        This property is not part of the returned xml of the configuration, but is set after
+        a call to `SoundTouchClient.GetPresetList(resolveSourceTitles=True)' so that source
+        titles can be displayed by user-interfaces.
+        """
+        return self._SourceTitle
+
+    @SourceTitle.setter
+    def SourceTitle(self, value:str):
+        """ 
+        Sets the SourceTitle property value.
+        """
+        self._SourceTitle = value
+
+
+    @property
     def TypeValue(self) -> str:
         """ Specifies the type of the content item. """
         if self._ContentItem is None:
@@ -250,6 +272,9 @@ class Preset(SoundTouchModelRequest):
         if self._ContentItem is not None:
             elmNode = self._ContentItem.ToElement()
             elm.append(elmNode)
+
+        if isRequestBody == False:
+            if self._SourceTitle is not None and len(self._SourceTitle) > 0: elm.set('SourceTitle', str(self._SourceTitle))
             
         return elm
 
@@ -263,6 +288,7 @@ class Preset(SoundTouchModelRequest):
         if self._ContentItem is not None: msg = '%s %s' % (msg, str(self._ContentItem))
         msg = '%s CreatedOn="%s"' % (msg, str(self._CreatedOn))
         msg = '%s UpdatedOn="%s"' % (msg, str(self._UpdatedOn))
+        if self._SourceTitle is not None: msg = '%s SourceTitle="%s"' % (msg, str(self._SourceTitle))
         return msg